Workspace view in jupyter (#409)
* WIP: Implement WorkspaceViewJp

* fix WorkspaceViewJp

* fix jupyter compile

* Add example2.ipynb

* fix memory issue on install test and misc updates

* fix github action

* trigger github action
